🐛 The regular expression for determining text that is italic (_) was greedy
One of the problems at GitHub that we saw when rending a markdown file
with some code blocks is that when there was an unclosed block, it will
highlight the rest of the remaining text as italic, overriding any
highlights that should take effect.

This test change is based on a refactoring that was done to group
together the patterns and also change the regex for bold to match the
one used in `burodepeper/language-markdown` which is also an atom
package for syntax highlighting

This test change also has a little conflicting test because `is*italic*`
stylizes the italic word while `not_italic_` does not. I think this is
expected behavior given the other uses of `_`. This is also ready the
behavior in GitHub production right now.

![prod-format](https://user-images.githubusercontent.com/1518902/83576959-d4a1de00-a4e7-11ea-802f-0ff7064a9309.png)