🎨 Refactor grammar file into groups for better readability
This change groups patterns into categories. It then composes these into
patterns that can be mixed and matched. This is not purely as asthetic
refactor, this addresses a limitation that was found by a GitHub
customer. They reported that if they had a `__` (double underscore),
which is a valid character in objective c,inside a code block, it will
override the rest of the file and highlight it as bold.

This change carries over some regular expressions for bold characters
from the `burodepeper/language-markdown` grammar which is also an atom
package. This new regular expression is not as _greedy_ as the previous
one so it does not override the rest of the file.

Please see the description of this PR for a more detailed explantion of
the process.